Fix unit-test
