+++ tests/flytekit/integration/remote/workflows/basic/dict_str_wf.py
@@ -4,7 +4,7 @@ import os
 
 from flytekit import task, workflow
 
-IMAGE = os.getenv("FLYTEKIT_IMAGE", "localhost:30000/flytekit:dev")
+IMAGE = os.environ.get("FLYTEKIT_IMAGE", "localhost:30000/flytekit:dev")
 
 
 @task(container_image=IMAGE)
@@ -18,4 +18,6 @@ def my_dict_str_wf(d: typing.Dict[str, str]) -> str:
 
 
 if __name__ == "__main__":
-    print(f"Running my_wf(d={{'a': 'rwx', 'b': '42'}})={my_dict_str_wf(d={'a': 'rwx', 'b': '42'})}")
+    print(
+        f"Running my_wf(d={{'a': 'rwx', 'b': '42'}})={my_dict_str_wf(d={'a': 'rwx', 'b': '42'})}"
+    )
